added color fadind and shading , case size, z_size

+++ src/rasterisation/fdf_pixel_put.c
@@ -6,7 +6,7 @@
 /*   By: bamrouch <bamrouch@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/31 11:45:52 by bamrouch          #+#    #+#             */
-/*   Updated: 2023/01/31 11:52:26 by bamrouch         ###   ########.fr       */
+/*   Updated: 2023/02/11 14:20:41 by bamrouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,6 @@ void	fdf_pixel_put(t_fdf *fdf, int x, int y, int color)
 	unsigned int	*dst;
 
 	dst = (unsigned int *) fdf->frame.addr;
-	if (x >= 0 && (unsigned int)x < fdf->window_info.width && y >= 0 && (unsigned int)y < fdf->window_info.height )
+	if (x >= 0 && x < fdf->window_info.width && y >= 0 && y < fdf->window_info.height )
 		dst[x + (y * fdf->window_info.width)] = (unsigned int) color;
 }
